requirements.txt
matplotlib
pyserial
customtkinter